backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
pillow
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
